logout - operate

+++ security/src/main/java/com/security/security/repository/TokenRepository.java
@@ -12,7 +12,7 @@ public interface TokenRepository extends JpaRepository<Token, Long> {
                 select t from Token t inner join User u  on t.user.id  = u.id
                 where u.id = :id and (t.expired = false or t.revoked = false)
             """)
-    List<Token> findAllValidTokensByUser(Long id);
+    List<Token> findAllValidTokensByUser(Integer id);
 
     Optional<Token> findByToken(String token);
 }
